C Server : Disable bonjour service
It turns out using bonjour service will trigger a allow local network permission dialog on the device. Even though, this dialog will show up only once after granting the permission, this could be a source of test failure if the app was uninstalled. As iPhone will always connect to the test machine, connect to the device using .local address (but not going through en0) is a better alternative.