Sync gateway put database payload now accepts full config files
It doesn't attempt to parse them any more than it needs to in order to function with the cloud class.  This means that the configs need to explicitly include num_index_replicas now since python will not add it.  The payloads are now read only.  To modify you will need to create a new payload with a modified dictionary.