CBL-4977 : Wait for all indexes removed before recreating SG database (#63)
The bucket's indexes will be deleted asynchronously after the bucket is dropped. When recreating the sg database, sg may wrongly detect that the indexes already exist, but later when trying to use the indexes for querying, the index-not-available error occurs as the index has already been deleted by that time.

Wait until all indexes are removed will help prevent the issue. It's important to wait after the bucket and its collections are created, otherwise, QueryIndexManager will not be able to return the pending-to-removed indexes created for the collections.