Overhaul logging in .NET server and fix an update database issue
Update database, in the presence of updated blobs, was adding blobs to every document instead of just the ones it was meant to.  Move to Serilog logging so that I can have a combined test server and cbl log in serialized order.  This requires integrating more heavily with MAUI, so give up on the headless server.