Fix test fest tests when running on real network
There is a flaw in the steps which says the following:

- start a push from db1 to sgw
- start a pull from sgw to db2
- wait for document events in db2

The third step is going to timeout because it doesn't account for the delay in step 1 when blobs are involved.  The pull is going to be idle until the entire blob finished and the document is actually recorded in sgw.