Add in a smelly short term solution for Java stop
The issue is that there is no information passed into the teardown about whether or not the test server was downloaded.  In the run tests phase it actually downloads the test server and marks it as downloaded as a side effect.  However, when stopping this doesn't happen so I added in a sneaky optional argument written in a way that I will continue to notice to both fix the problem and remind me to do a better fix later.