Uncomment dynamic branch selection for prebuild (#213)
This will attempt to select the appropriate build branch based on our naming convention and the version number passed.  I had to modify not using matcher directly in a variable because of a wildly unexpected quirk of Jenkins that tries to serialize the state of the pipeline after every action, resulting in a NotSerializableError because matchers are not serializable.