JS Server work
* JS Server: Implemented snapshot & verify

Also updated CBL dependency to 1.0.0-5

* JS Server: Implemented blobs for /updateDatabase

* JS Server: Normalize collection names everywhere

i.e. convert `_default._default` to `_default`

* JS Server: Actually save doc after updating it

* Disable more JS tests and a few more test server fixes

1. Make it an error to have a snapshot update to a document that doesn't exist in a snapshot
2. Only return result.document if the snapshot update type is an update (not delete or purge)
3. Sometimes we make a snapshot of a document that does not exist in the DB in order to verify that it didn't get pulled unintentionally, so check that (it will show up as an entry with an "undefined" value)
4. updateDatabase should be allowed to create documents
5. Bug in update remove properties handling (it was iterating the actual strings and sending them one character at a time)

* Fix API spec for conflict resolvers and run query test

* JS Server: Added auth support for replicator

* JS Server: Implemented conflict resolvers

* JS Server: Throw error when updating keypath fails

* JS Server: Implemented push/pull filters

* JS Server: Resolve relative blob URLs

* JS Server: support updatedBlobs in /verifyDocuments

* JS Server: Just prettified tdkSchema.ts

* JS Server: Fixed transaction error in /updateDatabase

Updating a blob would trigger an exception because it's illegal to
make non-database async calls within a transaction.
So use Collection.updateMultiple() instead.

NOTE: This assumes that a /updateDatabase request does not list the
same document multiple times! If so, this will probably fail.

* A few small fixes

1. Normalize collection name in replicator config
2. Correct blob base URL

* Correct snapshot tests and behavior

Only update verifications (not delete or purge) return the document.  Also, there is a distinction between null and missing for snapshot entries:  null means "I want to verify this later, but it doesn't exist right now".  undefined is "Not included in the snapshot and ineligible for verification"

* Further correct Snapshot

- Spurious `!` on line 40 messes up the test for undefined
- I accidentally used `!!` (Kotlin syntax) instead of `!` on line 59
- Using `T | null` in DocumentMap is incorrect; nulls are handled in
  the declaration of #documents, which is
  `DocumentMap<cbl.CBLDocument | null>`.

* Correct and error code and prettify message

The TDK expects an HTTP 400 when a nonexistent blob is requested.  Remove "self" from error message since it results in it being printed twice, and make the error message "returned XXX" not weirdly formatted.

* Update travel JS dataset

Needs two empty collections to be compatible with SGW's setup

* Switch SGW certificate strategy

The cert is now issued by a CA that can easily be trusted in python, as well as browser.  Existing SDK cert pinning should still be fine as well.  Also only copy prerelease RPM if it's not already on the remote machine (uploads are slow)

* Emit "transport: ws" for Javascript topology setups

* Need to separate push and pull filter

Their functions are different signatures so a premade CreateFilter is not possible

* Relax smoke test requirements for error on connection failure

Since JS has no access to that information :(

---------

Co-authored-by: Jim Borden <jim.borden@couchbase.com>