Update test_filter_removed_access_documents for JS
Not checking for the filtered access removed doc notification for JS. Based on the discussion, we will not implement document ended notification for the docs rejected by the pull filter in JS. We may implement it based on demand.

From Jens, this feature doesn’t make sense as the rejection from the user’s replication filters shouldn’t be treat as an error. Also it happens on the platform as a side effect due to an error thrown by LiteCore. I have created CBL-7645 for changing the behavior on the other platforms in the future release.