Switch to t4g instances where possible (#319)
Edge server is still x86_64 only so it uses t3.  Also add in queries to find the latest AMI for Amazon Linux 2023 for both arm and amd.  Download SGW aarch64 packages.

This requires some slight tweaking to how log slurp is built so that it can be a multi platform container.